requirements.txt
flask
torch
torchvision
opencv-python
ultralytics
pillow
numpy
scikit-learn
tqdm

